requirements/requirements_user_app_dj18.txt
Django==1.8.3
IPy==0.83
MySQL-python==1.2.5
Pillow==2.9.0
PyYAML==3.11
Shapely==1.5.9
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.20
celery==3.1.18
django-activity-stream==0.6.0
django-appconf==1.0.1
django-bootstrap-breadcrumbs==0.7.1
django-celery==3.1.16
django-cors-headers==1.1.0
django-datatables-view==1.12.1
django-debug-panel==0.8.1
django-debug-toolbar==1.3.2
django-debug-toolbar-line-profiler==0.4.0
django-extensions==1.5.5
django-jsonfield==0.9.13
django-jsonify==0.3.0
django-mptt==0.7.4
django-passwords==0.3.3
django-realtime==1.1
django-redis==4.2.0
django-redis-cache==1.1.1
django-rest-swagger==0.3.2
django-session-security==2.2.4
django-settings-export==1.0.5
django-staticfiles==1.2.1
djangorestframework==3.1.3
hiredis==0.2.0
ipaddr==2.1.11
kombu==3.0.26
line-profiler==1.0
memory-profiler==0.33
msgpack-python==0.4.6
pydot==1.0.2
pyenchant==1.6.6
pymongo==3.0.3
pyparsing==2.0.3
pyproj==1.9.4
python-dateutil==2.4.2
python-memcached==1.54
pytz==2015.4
raven==5.4.4
redis==2.10.3
requests==2.7.0
simplejson==3.7.3
six==1.9.0
sqlparse==0.1.15
uWSGI==2.0.11
ujson==1.33
wsgiref==0.1.2
xlrd==0.9.3
xlwt==1.0.0
